Move LocalChat to a seperated module
