+++ d2bs/kolbot/libs/common/Misc.js
@@ -2292,160 +2292,3 @@ function PacketBuilder () {
 	this.spoof = () => (getPacket(this.buildDataView().buffer), this);
 	this.get = this.spoof; // same thing but spoof has clearer intent than get
 };
-
-var LocalChat = new function () {
-	const LOCAL_CHAT_ID = 0xD2BAAAA;
-	let toggle, proxy = say;
-
-	let relay = (msg) => D2Bot.shoutGlobal(JSON.stringify({ msg: msg, realm: me.realm, charname: me.charname, gamename: me.gamename }), LOCAL_CHAT_ID);
-
-	let onChatInput = (speaker, msg) => {
-		relay(msg);
-		return true;
-	};
-
-	let onChatRecv = (mode, msg) => {
-		if (mode !== LOCAL_CHAT_ID) {
-			return;
-		}
-
-		msg = JSON.parse(msg);
-
-		if (me.gamename === msg.gamename && me.realm === msg.realm) {
-			new PacketBuilder().byte(38).byte(1, me.locale).word(2, 0, 0).byte(90).string(msg.charname, msg.msg).get();
-		}
-	};
-
-	let onKeyEvent = (key) => {
-		if (toggle === key) {
-			this.init(true);
-		}
-	};
-
-	this.init = (cycle = false) => {
-		if (!Misc.config.LocalChat.Enabled) {
-			return;
-		}
-
-		Misc.config.LocalChat.Mode = (Misc.config.LocalChat.Mode + cycle) % 3;
-		print("ÿc2LocalChat enabled. Mode: " + Misc.config.LocalChat.Mode);
-
-		switch (Misc.config.LocalChat.Mode) {
-		case 2:
-			removeEventListener("chatinputblocker", onChatInput);
-			addEventListener("chatinputblocker", onChatInput);
-		case 1:
-			removeEventListener("copydata", onChatRecv);
-			addEventListener("copydata", onChatRecv);
-			say = (msg, force = false) => force ? proxy(msg) : relay(msg);
-			break;
-		case 0:
-			removeEventListener("chatinputblocker", onChatInput);
-			removeEventListener("copydata", onChatRecv);
-			say = proxy;
-			break;
-		}
-
-		if (Misc.config.LocalChat.Toggle) {
-			toggle = typeof Misc.config.LocalChat.Toggle === 'string' ? Misc.config.LocalChat.Toggle.charCodeAt(0) : Misc.config.LocalChat.Toggle;
-			Misc.config.LocalChat.Toggle = false;
-			addEventListener("keyup", onKeyEvent);
-		}
-	};
-};
-
-var Messaging = {
-	sendToScript: function (name, msg) {
-		var script = getScript(name);
-
-		if (script && script.running) {
-			script.send(msg);
-
-			return true;
-		}
-
-		return false;
-	},
-
-	sendToProfile: function (profileName, mode, message, getResponse) {
-		var response;
-
-		function copyDataEvent(mode2, msg) {
-			if (mode2 === mode) {
-				var obj;
-
-				try {
-					obj = JSON.parse(msg);
-				} catch (e) {
-					return false;
-				}
-
-				if (obj.hasOwnProperty("sender") && obj.sender === profileName) {
-					response = Misc.copy(obj);
-				}
-
-				return true;
-			}
-
-			return false;
-		}
-
-		if (getResponse) {
-			addEventListener("copydata", copyDataEvent);
-		}
-
-		if (!sendCopyData(null, profileName, mode, JSON.stringify({message: message, sender: me.profile}))) {
-			//print("sendToProfile: failed to get response from " + profileName);
-
-			if (getResponse) {
-				removeEventListener("copydata", copyDataEvent);
-			}
-
-			return false;
-		}
-
-		if (getResponse) {
-			delay(200);
-			removeEventListener("copydata", copyDataEvent);
-
-			if (!!response) {
-				return response;
-			}
-
-			return false;
-		}
-
-		return true;
-	}
-};
-
-var Events = {
-	// gamepacket
-	gamePacket: function (bytes) {
-		var temp;
-
-		switch (bytes[0]) {
-		// Block movement after using TP/WP/Exit
-		case 0x0D: // Player Stop
-			// This can mess up death screen so disable for characters that are allowed to die
-			if (Misc.config.LifeChicken > 0) {
-				return true;
-			}
-
-			break;
-		// Block poison skills that might crash the client
-		case 0x4C: // Cast skill on target
-		case 0x4D: // Cast skill on coords
-			temp = Number("0x" + bytes[7].toString(16) + bytes[6].toString(16));
-
-			// Match Poison Javelin, Plague Javelin or Poison Nova
-			if (temp && [15, 25, 92].indexOf(temp) > -1) {
-				return true;
-			}
-
-			break;
-		}
-
-		return false;
-	}
-};
